+++ nucliadb_vectors/src/relations/service/reader.rs
@@ -84,13 +84,19 @@ impl RelationsReaderService {
         }
         bfs_request.node_filters.iter().for_each(|filter| {
             let node_type = filter.node_type();
-            let node_subtype = filter.node_subtype.as_ref().map_or_else(|| "", |subtype| subtype);
+            let node_subtype = filter
+                .node_subtype
+                .as_ref()
+                .map_or_else(|| "", |subtype| subtype);
             let type_info = node_type_parsing(node_type, node_subtype);
             node_filters.insert(type_info);
         });
         bfs_request.edge_filters.iter().for_each(|filter| {
             let relation_type = filter.relation_type();
-            let relation_subtype = filter.relation_subtype.as_ref().map_or_else(|| "", |subtype| subtype);
+            let relation_subtype = filter
+                .relation_subtype
+                .as_ref()
+                .map_or_else(|| "", |subtype| subtype);
             let type_info = relation_type_parsing(relation_type, relation_subtype);
             edge_filters.insert(type_info);
         });
@@ -136,7 +142,9 @@ impl RelationsReaderService {
         if let Ok(v) = time.elapsed().map(|s| s.as_millis()) {
             info!("{id:?} - Ending at {v} ms");
         }
-        Ok(Some(EntitiesSubgraphResponse { relations: subgraph }))
+        Ok(Some(EntitiesSubgraphResponse {
+            relations: subgraph,
+        }))
     }
     #[tracing::instrument(skip_all)]
     fn prefix_search(